Convert Application to use Spring Dependency Injection and use GameCacheInMemoryImpl
Converted the application to use Spring Dependency Injection to make life easier. This will allow us to scale the application easier and be less reliant on constructing objects within the TwoRoomsIoApplication class and let Spring take care of it.

I've also wired up the GameCache to the GameEngine for now. It's currently still a single game usage only but will fix Player app to start sending the Game ID for all it's actions so that we can properly move to multi-tenant games in the future.